Create Clock-specific assertion module
